Proyectos/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.7
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap4==23.2
django-colorfield==0.10.1
django-jazzmin==2.6.0
et-xmlfile==1.1.0
mysql-connector-python==8.2.0
Pillow==10.1.0
protobuf==4.21.12
pydantic==2.5.2
pydantic_core==2.14.5
python-slugify==8.0.1
reportlab==4.0.7
soupsieve==2.5
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.8.0
tzdata==2023.3
